dataset definition and augmentation experimenting with contrastive training
